refactor(shared-server): isolate client mutation transaction shell
