+++ packages/shared-server/mod.ts
@@ -21,7 +21,8 @@ export * from './rallar-system/repositories/RtcTopologyExecutionRepository.ts';
 export * from './rallar-system/repositories/StateEventStore.ts';
 export * from './rallar-system/pubsub/QueueBoxPubSubBridge.ts';
 export * from './rallar-system/pubsub/RtcTopologyClusterTransport.ts';
-export * from './rallar-system/services/client-state-service.ts';
+export * from './rallar-system/client-state/client-state-service-contracts.ts';
+export * from './rallar-system/client-state/client-state-service.ts';
 export * from './rallar-system/services/client-state-snapshot-read-through-cache.ts';
 export * from './rallar-system/services/cached-client-state-service.ts';
 export * from './rallar-system/services/cached-group-state-service.ts';
@@ -29,7 +30,8 @@ export * from './rallar-system/services/auth-login-service.ts';
 export * from './rallar-system/services/auth-state-mutations.ts';
 export * from './rallar-system/services/auth-credential-issuer.ts';
 export * from './rallar-system/services/AppAuthInboxService.ts';
-export * from './rallar-system/services/AppClientInboxService.ts';
+export * from './rallar-system/client-state/inbox/app-client-inbox-contracts.ts';
+export * from './rallar-system/client-state/inbox/app-client-inbox-service.ts';
 export * from './rallar-system/services/AppGroupInboxService.ts';
 export * from './rallar-system/services/AppInboxService.ts';
 export * from './rallar-system/services/AppOutboxService.ts';
