+++ plans/repo-style-lineages/client-state-server-ordinary-transaction.json
@@ -1,0 +1,32 @@
+{
+  "version": 1,
+  "lineages": [
+    {
+      "mergeBase": "2fdba024bb347622727d337eb06fc13d2fe129fc",
+      "source": {
+        "path": "packages/shared-server/rallar-system/services/client-state-service.ts",
+        "blob": "aa6c2483db49bfc2c819e14c37d64197a51064c7"
+      },
+      "targets": [
+        "packages/shared-server/rallar-system/client-state/client-state-service-contracts.ts",
+        "packages/shared-server/rallar-system/client-state/client-state-service.ts",
+        "packages/shared-server/rallar-system/client-state/client-state-service-timing.ts",
+        "packages/shared-server/rallar-system/client-state/mutation/read/read-client-mutation.ts",
+        "packages/shared-server/rallar-system/client-state/mutation/write/write-client-mutation.ts"
+      ]
+    },
+    {
+      "mergeBase": "2fdba024bb347622727d337eb06fc13d2fe129fc",
+      "source": {
+        "path": "packages/shared-server/rallar-system/services/AppClientInboxService.ts",
+        "blob": "8f5d371f3693e135e17beeeef4781aba19c93a23"
+      },
+      "targets": [
+        "packages/shared-server/rallar-system/client-state/inbox/app-client-inbox-contracts.ts",
+        "packages/shared-server/rallar-system/client-state/inbox/authenticated-client-mutation-ingress.ts",
+        "packages/shared-server/rallar-system/client-state/inbox/client-state-inbox-handler.ts",
+        "packages/shared-server/rallar-system/client-state/inbox/app-client-inbox-service.ts"
+      ]
+    }
+  ]
+}
